+++ lsp/initialize.go
@@ -83,7 +83,7 @@ func NewInitializeResponse(id int) InitializeResponse {
 			},
 			ServerInfo: ServerInfo{
 				Name:    "gsclsp",
-				Version: "0.0.7.4",
+				Version: "0.0.7.5",
 			},
 		},
 	}